requirements.txt
streamlit
google-generativeai>=0.5.0
beautifulsoup4
requests
